Promote all MediaNoche code except typematcher function.
